requirements.txt
-i https://pypi.org/simple
cachecontrol==0.12.11; python_full_version >= '3.6.0'
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.9.24; python_full_version >= '3.6.0'
charset-normalizer==2.1.1; python_full_version >= '3.6.0'
firebase-admin==5.4.0
google-api-core[grpc]==2.10.1; platform_python_implementation != 'PyPy'
google-api-python-client==2.63.0; python_version >= '3.7'
google-auth==2.12.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2; python_version >= '3.7'
google-cloud-firestore==2.7.0; platform_python_implementation != 'PyPy'
google-cloud-storage==2.5.0; python_version >= '3.7'
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.4.0; python_version >= '3.7'
googleapis-common-protos==1.56.4; python_version >= '3.7'
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4; python_version >= '3.5'
msgpack==1.0.4
proto-plus==1.22.1; python_full_version >= '3.6.0'
protobuf==4.21.7; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9; python_version > '3.0'
python-dotenv==0.21.0
requests==2.28.1; python_version >= '3.7' and python_version < '4'
rsa==4.9; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
uritemplate==4.1.1; python_full_version >= '3.6.0'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
